feat: i18n scripts
