+++ packages/scripts/README.md
@@ -7,3 +7,10 @@
 * test      - run tests (jest)
 
 [Go to documentation](./webapp/README.md)
+
+## I18n
+* i18n-extract  - extract the translation files from your project
+* i18n-upload   - upload extracted files to XTM
+* i18n-download - download files from XTM
+
+[Go to documentation](./i18n/README.md)
