+++ packages/scripts/package.json
@@ -17,10 +17,12 @@
     "@babel/preset-react": "^7.0.0",
     "@talend/html-webpack-plugin": "1.0.0",
     "@talend/react-cmf-webpack-plugin": "1.10.0-0",
+    "adm-zip": "0.4.10",
     "autoprefixer": "^7.1.4",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^23.6.0",
     "babel-loader": "^8.0.0",
+    "chalk": "^2.4.1",
     "copy-webpack-plugin": "^4.6.0",
     "cross-spawn": "^6.0.4",
     "css-loader": "^1.0.1",
@@ -34,11 +36,14 @@
     "eslint-plugin-jsx-a11y": "^2.2.2",
     "eslint-plugin-react": "^6.3.0",
     "file-loader": "^2.0.0",
+    "form-data": "^2.3.3",
     "html-webpack-plugin": "3.2.0",
     "jest": "^23.6.0",
     "lodash.get": "^4.4.2",
     "lodash.mergewith": "^4.6.1",
     "mini-css-extract-plugin": "^0.4.0",
+    "mkdirp": "^0.5.1",
+    "node-fetch": "^2.2.0",
     "node-sass": "^4.7.2",
     "optimize-css-assets-webpack-plugin": "4.0.0",
     "postcss-loader": "^3.0.0",
@@ -47,6 +52,7 @@
     "resolve-url-loader": "2.3.0",
     "rimraf": "^2.6.1",
     "sass-loader": "^7.1.0",
+    "shelljs": "^0.8.2",
     "uglifyjs-webpack-plugin": "1.2.7",
     "url-loader": "^1.1.2",
     "webpack": "^4.19.0",
@@ -55,6 +61,7 @@
     "whatwg-fetch": "^2.0.3",
     "which": "^1.3.0",
     "yargs-parser": "^9.0.2"
+
   },
   "peerDependencies": {
     "@talend/bootstrap-theme": "1.10.0",
